requirements.txt
python-dotenv
langchain
langchain-groq
langchain-huggingface
langchain-chroma
langchain-community
sentence-transformers
pypdf
streamlit==1.43.2
protobuf==4.25.3